feat(CoreData): change coredata modeling
